Move test factories into tests module
